# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Express App API",
      "description": "API documentation for the Express application",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:3000"
      }
    ],
    "paths": {
      "/": {
        "get": {
          "summary": "Root endpoint",
          "responses": {
            "200": {
              "description": "A successful response",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "string",
                    "example": "Hello world!"
                  }
                }
              }
            }
          }
        }
      },
      "/upload-csv": {
        "post": {
          "summary": "Upload CSV file",
          "requestBody": {
            "required": true,
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "csvfile": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "202": {
              "description": "Data received and saved to database"
            },
            "400": {
              "description": "No file uploaded or format error in CSV file",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Failed to save data to database",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/search": {
        "get": {
          "summary": "Search AGB codes",
          "parameters": [
            {
              "in": "query",
              "name": "q",
              "schema": {
                "type": "string"
              },
              "required": true,
              "description": "Search query"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful search operation",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/AgbCode"
                    }
                  }
                }
              }
            },
            "403": {
              "description": "Not permitted for search or unauthorized",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/getProfile/{id}": {
        "get": {
          "summary": "Get user profile",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "schema": {
                "type": "string"
              },
              "required": true,
              "description": "User ID"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful profile retrieval",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/AgbCode"
                  }
                }
              }
            },
            "403": {
              "description": "Not permitted or unauthorized to access the profile",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "description": "Profile not found",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/login": {
        "post": {
          "summary": "Login endpoint",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful login",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "token": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Invalid username or password",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "AgbCode": {
          "type": "object",
          "properties": {
            "agbCode": {
              "type": "string"
            },
            "name": {
              "type": "string"
            },
            "phoneNumber": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "street": {
              "type": "string"
            },
            "houseNumber": {
              "type": "string"
            },
            "houseNumberAddition": {
              "type": "string"
            },
            "postalCode": {
              "type": "string"
            },
            "city": {
              "type": "string"
            }
          }
        }
      }
    }
  }
  